Breaking up mangaInfos duck
