+++ src/components/FavoriteFab.js
@@ -7,9 +7,9 @@ import { makeStyles } from "@material-ui/styles";
 import { useSelector, useDispatch } from "react-redux";
 import {
   selectIsFavoriteToggling,
-  selectIsFavorite,
-  toggleFavorite
+  selectIsFavorite
 } from "redux-ducks/mangaInfos";
+import { toggleFavorite } from "redux-ducks/mangaInfos/actionCreators";
 
 // TODO: Loading spinner flickers because of short delay.
 //       Would be interesting to create a spinner with a small delay before appearing.
