+++ src/redux-ducks/catalogue.js
@@ -5,11 +5,8 @@ import type { FilterAnyType } from "types/filters";
 import type { MangaType } from "types";
 import { createLoadingSelector } from "redux-ducks/loading";
 import { createSelector } from "reselect";
-import {
-  selectMangaInfos,
-  ADD_MANGA,
-  type AddMangaAction
-} from "redux-ducks/mangaInfos";
+import { selectMangaInfos, type AddMangaAction } from "redux-ducks/mangaInfos";
+import type { ADD_MANGA } from "redux-ducks/mangaInfos/actions";
 import { selectLastUsedFilters } from "redux-ducks/filters";
 import { handleHTMLError, transformToMangaIdsArray } from "redux-ducks/utils";
